+++ plugin-dev/Source/Sentry/Private/SentrySettings.cpp
@@ -44,6 +44,9 @@ USentrySettings::USentrySettings(const FObjectInitializer& ObjectInitializer)
 	, UseProxy(false)
 	, ProxyUrl()
 	, DatabaseLocation(ESentryDatabaseLocation::ProjectUserDirectory)
+	, UseNativeBackend(false)
+	, MinidumpMode(ESentryMinidumpMode::Smart)
+	, CrashReportingMode(ESentryCrashReportingMode::NativeStackwalkingWithMinidump)
 	, CrashpadWaitForUpload(false)
 	, EnableOnCrashLogging(false)
 	, EnableExternalCrashReporter(false)
@@ -97,6 +100,7 @@ void USentrySettings::PostEditChangeProperty(FPropertyChangedEvent& PropertyChan
 	}
 
 	if (PropertyChangedEvent.Property->GetFName() == GET_MEMBER_NAME_CHECKED(USentrySettings, InitAutomatically) ||
+		PropertyChangedEvent.Property->GetFName() == GET_MEMBER_NAME_CHECKED(USentrySettings, UseNativeBackend) ||
 		PropertyChangedEvent.Property->GetFName() == GET_MEMBER_NAME_CHECKED(USentrySettings, UploadSymbolsAutomatically) ||
 		PropertyChangedEvent.Property->GetFName() == GET_MEMBER_NAME_CHECKED(USentrySettings, ProjectName) ||
 		PropertyChangedEvent.Property->GetFName() == GET_MEMBER_NAME_CHECKED(USentrySettings, OrgName) ||