+++ sample/Source/SentryPlayground/SentryPlaygroundGameInstance.cpp
@@ -18,6 +18,7 @@
 #include "SentryVariant.h"
 
 #include "CoreGlobals.h"
+#include "SentryModule.h"
 #include "HAL/Platform.h"
 #include "Misc/CoreDelegates.h"
 #include "Misc/EngineVersionComparison.h"
@@ -137,7 +138,16 @@ void USentryPlaygroundGameInstance::RunCrashTest(ESentryAppTerminationType Crash
 
 	// Workaround for duplicated log messages in UE 4.27 on Linux
 #if PLATFORM_LINUX && UE_VERSION_OLDER_THAN(5, 0, 0)
-	UE_LOG(LogSentrySample, Log, TEXT("EVENT_CAPTURED: %s\n"), *EventId);
+	if (FSentryModule::Get().GetSettings()->UseNativeBackend)
+	{
+		// When using native backend, crashing process can terminate before engine’s GLog is flushed
+		// To ensure the event ID is not lost it is printed directly to stdout so the test script can reliably capture it
+		FPlatformMisc::LowLevelOutputDebugStringf(TEXT("EVENT_CAPTURED: %s\n"), *EventId);
+	}
+	else
+	{
+		UE_LOG(LogSentrySample, Log, TEXT("EVENT_CAPTURED: %s\n"), *EventId);
+	}
 #else
 	UE_LOG(LogSentrySample, Display, TEXT("EVENT_CAPTURED: %s\n"), *EventId);
 #endif