+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add experimental native backend support for Windows and Linux ([#1295](https://github.com/getsentry/sentry-unreal/pull/1295))
 - Add automatic frame time metrics ([#1293](https://github.com/getsentry/sentry-unreal/pull/1293))
 
 ### Dependencies